Cambios sobre review y mejoras
