+++ .gitignore
@@ -1,7 +1,6 @@
 # Byte-compiled / optimized / DLL files
-__pycache__/
-*.py[cod]
-*$py.class
+*.py[co]
+.~*
 
 # C extensions
 *.so
