+++ sysadmin_base/README.rst
@@ -15,11 +15,23 @@ Implemented Uses
 **Stock**
 
 * Any product have an option of being an internal equipment, giving the posibility of a new page of Technical Information of the equipment.
-* Implements an option that allows you to populate hardware fields of products associated with a category that has the "Show hardware properties" option enabled.
-* You can also filter by "internal equipment".
+* Implements an option that allows you to populate hardware fields of items associated to products that are associated with a category that has the "Show hardware properties" option enabled.
+* Each product can be associated with an item.
 
 **Employees**
 
 * The employees have 2 fields that indicates if that employee has access to Dialoga and if that employee is trainee.
-* On tree view, you can see if the employees "Has Dialoga access".
-* Allows you to filter by "Has Dialoga acess" and group by "Is trainee".
+* Every employee has workspaces and items associated to it.
+
+**Workspaces**
+
+* Each workspace has employees and items asigned to it. It represents a real workspace.
+
+**Items**
+
+* Every item can be associated with an employee or a workspace.
+* Every item can be associated with a subsidy.
+
+**Subsidies**
+
+* Every subsidy can be associated with many items.
