+++ sysadmin_base/models/__init__.py
@@ -4,4 +4,6 @@
 from . import hr_employee
 from . import product_category
 from . import product_template
-from . import workspace
+from . import workspace_item
+from . import workspace_item_subsidy
+from . import workspace_workspace
