+++ sysadmin_base/security/sysadmin_security.xml
@@ -1,13 +1,16 @@
 <?xml version='1.0' encoding='utf-8'?>
 <odoo>
-    <data noupdate="0">
-        <record id="group_sysadmin_sysadmin" model="res.groups">
-            <field name="name">SysAdmin / SysAdmin</field>
-            <field name="implied_ids" eval="[(4, ref('base.group_user'))]"/>
-        </record>
-        <record id="group_sysadmin_employee" model="res.groups">
-            <field name="name">SysAdmin / Employee</field>
-            <field name="implied_ids" eval="[(4, ref('base.group_user'))]"/>
-        </record>
-    </data>
+
+    <record id="group_sysadmin_sysadmin" model="res.groups">
+        <field name="name">SysAdmin</field>
+        <field name="category_id" ref="base.module_category_sysadmin"/>
+        <field name="comment">The user will have access to all configuration of SysAdmin.</field>
+    </record>
+
+    <record id="group_sysadmin_employee" model="res.groups">
+        <field name="name">Employee</field>
+        <field name="category_id" ref="base.module_category_sysadmin"/>
+        <field name="comment">The user will be able to move it's items between it's workspaces.</field>
+    </record>
+
 </odoo>
