+++ sysadmin_base/views/hr_employee_views.xml
@@ -13,9 +13,6 @@
                         <field name="is_trainee"/>
                     </group>
                     <notebook>
-                        <page name="employee_workspaces" string="Workspaces">
-                            <field name="workspace_ids" context="{'default_employee_id': active_id}"/>
-                        </page>
                         <page name="employee_items" string="Items">
                             <field name="item_ids" context="{'default_employee_id': active_id}">
                                 <tree>
@@ -25,6 +22,17 @@
                                     <field name="workspace_id"/>
                                     <field name="workspace_location"/>
                                     <field name="subsidy_id"/>
+                                </tree>
+                            </field>
+                        </page>
+                        <page name="employee_workspaces" string="Workspaces">
+                            <field name="workspace_ids" context="{'default_employee_id': active_id}">
+                                <tree>
+                                    <field name="name"/>
+                                    <field name="description"/>
+                                    <field name="location"/>
+                                    <field name="item_ids"/>
+                                    <field name="employee_ids"/>
                                 </tree>
                             </field>
                         </page>
@@ -41,6 +49,7 @@
         <field name="arch" type="xml">
             <field name="work_phone" position="after">
                 <field name="has_dialoga_access"/>
+                <field name="is_trainee"/>
             </field>
         </field>
     </record>
