+++ sysadmin_base/views/workspace_item_subsidy_views.xml
@@ -1,0 +1,45 @@
+<?xml version='1.0' encoding='utf-8'?>
+<odoo>
+    
+    <record id="workspace_subsidy_view_form" model="ir.ui.view">
+        <field name="model">workspace.item.subsidy</field>
+        <field name="arch" type="xml">
+            <form string="Subsidy">
+                <sheet>
+                    <h1>
+                        <field name="name"/>
+                    </h1>
+                    <group>
+                        <field name="description"/>
+                        <field name="date"/>
+                    </group>
+                    <group>
+                        <field name="item_ids" context="{'default_subsidy_id': active_id}">
+                            <tree>
+                                <field name="name"/>
+                                <field name="sn"/>
+                                <field name="product_id"/>
+                                <field name="workspace_id"/>
+                                <field name="workspace_location"/>
+                                <field name="employee_id"/>
+                            </tree>
+                        </field>
+                    </group>
+                </sheet>
+            </form>
+        </field>
+    </record>
+
+    <record id="workspace_subsidy_view_tree" model="ir.ui.view">
+        <field name="model">workspace.item.subsidy</field>
+        <field name="arch" type="xml">
+            <tree>
+                <field name="name"/>
+                <field name="description"/>
+                <field name="date"/>
+                <field name="item_ids"/>
+            </tree>
+        </field>
+    </record>
+    
+</odoo>
