+++ sysadmin_base/views/workspace_item_views.xml
@@ -1,0 +1,96 @@
+<?xml version='1.0' encoding='utf-8'?>
+<odoo>
+    <record id="workspace_item_view_search" model="ir.ui.view">
+        <field name="model">workspace.item</field>
+        <field name="arch" type="xml">
+            <search string="Items">
+                <filter string="Workspaces" name="workspace_items" context="{'group_by':'workspace_id'}"/>
+                <filter string="Products" name="product_items" context="{'group_by':'product_id'}"/>
+                <filter string="Employee" name="employee_items" context="{'group_by':'employee_id'}"/>
+            </search>
+        </field>
+    </record>
+
+    <record id="workspace_item_view_kanban" model="ir.ui.view">
+        <field name="model">workspace.item</field>
+        <field name="arch" type="xml">
+            <kanban string="Items" default_group_by="workspace_id">
+                <field name="name"/>
+                <field name="sn"/>
+                <field name="workspace_location"/>
+                <templates>
+                    <t t-name="kanban-box">
+                        <div class="oe_kanban_global_click">
+                            <div t-attf-class="oe_kanban_content">
+                                <h3>
+                                    <field name="name"/>
+                                </h3>
+                                <b>SN:</b>
+                                <field name="sn"/>
+                                <br/>
+                                <b>Workspace location:</b>
+                                <field name="workspace_location"/>
+                            </div>
+                        </div>
+                    </t>
+                </templates>
+            </kanban>
+        </field>
+    </record>
+
+    <record id="workspace_item_view_tree" model="ir.ui.view">
+        <field name="model">workspace.item</field>
+        <field name="arch" type="xml">
+            <tree string="Item">
+                <field name="name"/>
+                <field name="description"/>
+                <field name="sn"/>
+                <field name="product_id"/>
+                <field name="workspace_id"/>
+                <field name="workspace_location"/>
+                <field name="employee_id"/>
+                <field name="subsidy_id"/>
+                <field name="cpu"/>
+                <field name="ram"/>
+                <field name="data_storage"/>
+                <field name="ip"/>
+                <field name="os_version"/>
+                <field name="microsoft_office_mail"/>
+            </tree>
+        </field>
+    </record>
+
+    <record id="workspace_item_view_form" model="ir.ui.view">
+        <field name="model">workspace.item</field>
+        <field name="arch" type="xml">
+            <form string="Item">
+                <sheet>
+                    <h1>
+                        <field name="name"/>
+                    </h1>
+                    <group>
+                        <group string="Information">
+                            <field name="description"/>
+                            <field name="sn"/>
+                            <field name="product_id"/>
+                            <field name="workspace_id"/>
+                            <field name="workspace_location"/>
+                            <field name="employee_id"/>
+                            <field name="subsidy_id"/>
+                        </group>
+                        <field name="hardware_properties" invisible="1"/>
+                        <group name="show_hardware_properties" string="Hardware Properties" attrs="{'invisible':[('hardware_properties','=',False)]}">
+                            <field name="cpu" placeholder="iX-XXXXNN / AMD XXXXX X / etc"/>
+                            <field name="ram"/>
+                            <field name="data_storage" placeholder="1TB-HDD / 512GB-SSD"/>
+                            <field name="ip" placeholder="X.X.X.X"/>
+                            <field name="os_version"/>
+                            <field name="microsoft_office_mail"/>
+                        </group>
+                    </group>
+                </sheet>
+            </form>
+        </field>
+    </record>
+
+</odoo>
